requirements.txt
Django==1.8.2
Markdown==2.6.2
-e git://github.com/stephenmcd/mezzanine.git#egg=Mezzanine
Pillow==2.8.2
SQLAlchemy==1.0.5
Werkzeug==0.10.4
beautifulsoup4==4.3.2
bleach==1.4.1
chardet==2.3.0
diff-match-patch==20110725.1
dj-database-url==0.3.0
django-appconf==1.0.1
django-compressor==1.5
django-contrib-comments==1.6.1
django-csvimport==2.4
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-import-export==0.2.7
django-postgrespool==0.3.0
filebrowser-safe==0.3.8
future==0.14.3
grappelli-safe==0.3.13
gunicorn==19.3.0
html5lib==0.99999
mezzanine-pagedown==0.8
oauthlib==0.7.2
psycopg2==2.6.1
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.15
tablib==0.10.0
tzlocal==1.2
whitenoise==2.0
wsgiref==0.1.2
